Fix post-payment redirect url when not installed in the root
In this case, the path to the installation below the root of the domain
is added by get_home_url, but also present in `REQUEST_URI`, so it ends
up in the resulting url twice.

Using `get_page_link()` just returns the url of the current page, which
is exactly what is needed here.